Add missing documentation for "lazyLoad"
I think it'd be good to add the lazy loading documentation for the `mapbox-gl` module. I've spent a few hours understanding how did it work because there was no documentation. I kept adding the `lazyLoad` flag under `ENV['mapbox-gl']` on the `config/environment.js` 🤦‍♂️

After I understood that it should be on the `ember-cli-build.js`, everything was easy.